* map special enchantments to itemenums to ease tracking * fix first two hp bars positions (by 1px to right) * fix alert indicators positions (by 1px to top) * fix close panel position in rings window (by 1px to right) * fix game_ui_minimap_frame not drawing * fix tooltip text alignment and vertical borders * fix opening multiple popups when two elements are close together without a gap * fix weapon recovery calculation * fix skills showing with bonuses in skill tab * fix awards scrollbar buttons crash * fix new game crash * fix pickup intersection * new pcx decoder: fix evil UI bottom panel, load pcx images in A8R8G8B8 color format, except for directdraw render for which is still use R5G6B5 * refactoring of some functions * nuklear: starting preparation for introducing party management debug menu * nuklear: fix not set window type in context

+++ Engine/Graphics/Vis.cpp
@@ -634,7 +634,7 @@ bool Vis::CheckIntersectBModel(BLVFace *pFace, Vec3_short_ IntersectPoint, signe
                 if (!v10) {
                     ++v16;
                 } else {
-                    float _v1 = ((intersect_face_vertex_coords_list_a[i + 1] - intersect_face_vertex_coords_list_a[i])) / (intersect_face_vertex_coords_list_b[i + 1] - intersect_face_vertex_coords_list_b[i]);
+                    float _v1 = (float)((intersect_face_vertex_coords_list_a[i + 1] - intersect_face_vertex_coords_list_a[i])) / (float)(intersect_face_vertex_coords_list_b[i + 1] - intersect_face_vertex_coords_list_b[i]);
                     int _v2 = (b - intersect_face_vertex_coords_list_b[i]) * _v1 + 0.5;
 
                     if (intersect_face_vertex_coords_list_a[i] + (_v2) >= a)
